% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R
\docType{methods}
\name{dbBind,KineticaResult-method}
\alias{dbBind,KineticaResult-method}
\title{dbBind()}
\usage{
\S4method{dbBind}{KineticaResult}(res, params, ...)
}
\arguments{
\item{res}{an object of [KineticaResult-class]}

\item{params}{A list of bindings, named or unnamed.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Binds the prepared statement to parameter values provided
}
\seealso{
Other KineticaResult methods: \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbFetch}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaResult methods}
