% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R
\docType{methods}
\name{dbClearResult,KineticaResult-method}
\alias{dbClearResult,KineticaResult-method}
\title{dbClearResult()}
\usage{
\S4method{dbClearResult}{KineticaResult}(res, ...)
}
\arguments{
\item{res}{[KineticaResult-class]}

\item{...}{Other parameters passed on to methods.}
}
\description{
Clears result, if exists, releasing memory in Connection results environment
}
\seealso{
Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbFetch}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaResult methods}
