% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R
\docType{methods}
\name{dbColumnInfo}
\alias{dbColumnInfo}
\alias{dbColumnInfo,KineticaResult-method}
\title{dbColumnInfo()}
\usage{
dbColumnInfo(res, ...)

\S4method{dbColumnInfo}{KineticaResult}(res, ...)
}
\arguments{
\item{res}{An object inheriting from [KineticaResult-class]}

\item{...}{Other parameters passed on to methods.}

\item{res}{An object inheriting from [KineticaResult-class]}

\item{...}{Other parameters passed on to methods.}
}
\description{
Provides data.frame as a collection of column names and data types
from the result set

Provides data.frame as a collection of column names and data types
from the result set
}
\seealso{
Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbFetch}}, \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}

Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbFetch}}, \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaResult methods}
