% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbConnect,KineticaDriver-method}
\alias{dbConnect,KineticaDriver-method}
\title{dbConnect()}
\usage{
\S4method{dbConnect}{KineticaDriver}(drv, host = NULL, port = NULL,
  url = NULL, username = NULL, password = NULL, timeout = NULL,
  ...)
}
\arguments{
\item{drv}{Kinetica driver object}

\item{host}{character string for Kinetica DB host address}

\item{port}{integer value for Kinetica DB port}

\item{url}{character string for Kinetica DB url (protocol + host + port)}

\item{username}{character string for Kinetica DB username}

\item{password}{character string for Kinetica DB password}

\item{timeout}{integer value for Kinetica DB connection timeout}

\item{...}{Other arguments ommited in generic signature}
}
\description{
Connect to a Kinetica DB going through the appropriate authentication procedure.
You can have multiple connections open and connections can have multiple results attached.
This function may be invoked repeatedly assigning its output to different
objects.
Authentication can be provided by using username/password variables.
Use [dbCanConnect()] to check if a connection can be established.
}
\examples{
\dontrun{
# If you jave a Kinetica DB instance running in local docker image,
# the minimum set of parameters to connect is url
dbConnect(Kinetica(), url = "http://localhost:9191")
# You can also use host and port combination
dbConnect(Kinetica(), host = "127.0.0.1", port = 9191)
# Remote instances would also require username and password.
dbConnect(Kinetica(), host = "127.0.0.1", port = 9191,
  username = "username", password = "Pa$$w0rd")
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
show(con)
# KineticaConnection
# url: http://localhost:9191
dbDisconnect(con)
}
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbGetInfo}}, \code{\link{dbIsValid}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
