% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbDataType,KineticaDriver-method}
\alias{dbDataType,KineticaDriver-method}
\title{Returns a matching SQLtype used when creating Kinetica DB object for the R data object}
\usage{
\S4method{dbDataType}{KineticaDriver}(dbObj, obj, ...)
}
\arguments{
\item{obj}{data object}
}
\description{
Returns a matching SQLtype used when creating Kinetica DB object for the R data object
}
\examples{
\dontrun{
dbDataType(Kinetica(), 1L)
> [1] "INTEGER"
dbDataType(Kinetica(), 1)
> [1] "DOUBLE"
dbDataType(Kinetica(), "Test string")
> [1] "VARCHAR(256)"
dbDataType(Kinetica(), Sys.Date())
> [1] "DATE"
dbDataType(Kinetica(), list())
> [1] "BINARY"
}
}
