% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbDisconnect,KineticaConnection-method}
\alias{dbDisconnect,KineticaConnection-method}
\title{Disconnects the current connection, releasing memory from all uncleared result objects}
\usage{
\S4method{dbDisconnect}{KineticaConnection}(conn, ...)
}
\description{
Disconnects the current connection, releasing memory from all uncleared result objects
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbDisconnect(con)
}
}
