% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbExecute,KineticaConnection,character-method}
\alias{dbExecute,KineticaConnection,character-method}
\title{Executes the statement, returning the number of affected rows or objects}
\usage{
\S4method{dbExecute}{KineticaConnection,character}(conn, statement, ...,
  params = NULL)
}
\arguments{
\item{statement}{character}

\item{params}{a set of values}
}
\description{
Executes the statement, returning the number of affected rows or objects
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbCreateTable(con, "tableA", data.frame(a = 1))
dbExistsTable(con, "tableA")
> [1] TRUE
dbExecute(con, "DROP TABLE tableA")
> [1] 1
dbDisconnect(con)
}
}
