% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbExistsTable,KineticaConnection,character-method}
\alias{dbExistsTable,KineticaConnection,character-method}
\title{Checks if the table exists.
Provide either table name or collection name, not both
Kinetica DB table names are unique, different collections don't allow tables with the same name.}
\usage{
\S4method{dbExistsTable}{KineticaConnection,character}(conn, name, ...)
}
\arguments{
\item{name}{character}
}
\description{
Checks if the table exists.
Provide either table name or collection name, not both
Kinetica DB table names are unique, different collections don't allow tables with the same name.
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbCreateTable(con, "tableA", data.frame(a = 1))
dbExistsTable(con, "tableA")
> [1] TRUE
dbExecute(con, "DROP TABLE tableA")
> [1] 1
dbExistsTable(con, "tableA")
> [1] FALSE
dbDisconnect(con)
}
}
