% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\name{dbExistsTable}
\alias{dbExistsTable}
\title{dbExistsTable()}
\usage{
dbExistsTable(conn, name, ...)
}
\arguments{
\item{conn}{an object of [KineticaConnection-class]}

\item{name}{character string for table or collection name  or [KineticaId-class] object}

\item{...}{Other parameters passed on to methods.}
}
\value{
logical
}
\description{
Checks if the table exists.
Provide either table name or collection name, not both
Kinetica DB table names are unique, different collections don't allow tables with the same name.
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbCreateTable(con, "tableA", data.frame(a = 1))
dbExistsTable(con, "tableA")
# TRUE
dbExecute(con, "DROP TABLE tableA")
# 1
dbExistsTable(con, "tableA")
# FALSE
dbDisconnect(con)
}
}
\seealso{
Other KineticaConnection methods: \code{\link{dbAppendTable}},
  \code{\link{dbCreateTable}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid}}, \code{\link{dbListFields}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}},
  \code{\link{transactions}}
}
\concept{KineticaConnection methods}
