% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R
\docType{methods}
\name{dbFetch}
\alias{dbFetch}
\alias{dbFetch,KineticaResult-method}
\title{dbFetch()}
\usage{
dbFetch(res, n = -1, ...)

\S4method{dbFetch}{KineticaResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from [KineticaResult-class]}

\item{n}{maximum number of records to retrieve per fetch. Use `n = -1`
or `n = Inf`
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}

\item{res}{An object inheriting from [KineticaResult-class]}

\item{n}{maximum number of records to retrieve per fetch. Use `n = -1`
or `n = Inf`
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Fetches the provided number of rows from Result. If n is not provided,
returns all available records

Fetches the provided number of rows from Result. If n is not provided,
returns all available records
}
\seealso{
Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}

Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaResult methods}
