% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbGetInfo,KineticaConnection-method}
\alias{dbGetInfo,KineticaConnection-method}
\title{Basic properties of KineticaConnection object
provides DB version, url, host, port and username of the current connection}
\usage{
\S4method{dbGetInfo}{KineticaConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{object [KineticaConnection-class]}
}
\description{
Basic properties of KineticaConnection object
provides DB version, url, host, port and username of the current connection
}
\examples{
\dontrun{
con <- dbConnect(d, url = "http://localhost:9191")
> $db.version
> [1] "7.0.0.0"
> $dbname
> [1] "Kinetica"
> $url
> [1] "http://localhost:9191"
> $host
> [1] "localhost"
> $port
> [1] 9191
> $username
> [1] ""
}
}
