% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbGetInfo,KineticaDriver-method}
\alias{dbGetInfo,KineticaDriver-method}
\title{Retrieves brief Driver information, dbname and driver/client version}
\usage{
\S4method{dbGetInfo}{KineticaDriver}(dbObj)
}
\arguments{
\item{dbObj}{an object derived of [KineticaDriver-class] type}
}
\description{
Retrieves brief Driver information, dbname and driver/client version
}
\examples{
\dontrun{
# Retrieves brief Driver information
drv <- Kinetica()
dbGetInfo(drv)
}
}
