% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaObject.R, R/kineticaDriver.R,
%   R/kineticaConnection.R, R/kineticaResult.R
\docType{methods}
\name{dbGetInfo}
\alias{dbGetInfo}
\alias{dbGetInfo,KineticaDriver-method}
\alias{dbGetInfo,KineticaConnection-method}
\alias{dbGetInfo,KineticaResult-method}
\title{dbGetInfo()}
\usage{
dbGetInfo(dbObj, ...)

\S4method{dbGetInfo}{KineticaDriver}(dbObj, ...)

\S4method{dbGetInfo}{KineticaConnection}(dbObj, ...)

\S4method{dbGetInfo}{KineticaResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from [KineticaObject-class]}

\item{...}{Other arguments to methods.}

\item{dbObj}{an object derived of [KineticaDriver-class] type}

\item{...}{Other arguments ommited in generic signature}

\item{dbObj}{object [KineticaConnection-class]}

\item{...}{Other arguments ommited in generic signature}

\item{dbObj}{[KineticaResult-class]}

\item{...}{Other parameters passed on to methods.}
}
\value{
a named list
}
\description{
Generic dbGetInfo() implementation.

Retrieves brief KineticaDriver description:
dbname, driver version, client version, max number of connections
(NA stands for unlimited)

Basic properties of KineticaConnection object
provides DB version, primary url, host, port and username of the
current connection, flag for HA availability, HA ring of secondary
urls and maximum row count that can be returned in one query

Provides basic info on KineticaResult object
}
\examples{
\donttest{
drv <- Kinetica()
dbGetInfo(drv)
# name
#  "KineticaDriver"
# dbname
#  "Kinetica"
# driver.version
#  "7.0.0.0"
# client.version
#  "7.0.0.0"
# max.connections
#  NA
}
\dontrun{
con <- dbConnect(d, url = "http://localhost:9191")
# db.version
# "7.0.0.0"
# dbname
# "Kinetica"
# url
# "http://localhost:9191"
# host
# "localhost"
# port
# 9191
# username
# ""
}
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbIsValid}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}

Other KineticaConnection methods: \code{\link{dbAppendTable,KineticaConnection-method}},
  \code{\link{dbCreateTable,KineticaConnection-method}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbExistsTable,KineticaConnection,character-method}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid,KineticaConnection-method}},
  \code{\link{dbListFields,KineticaConnection,character-method}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable,KineticaConnection,character-method}},
  \code{\link{dbRemoveTable,KineticaConnection,character-method}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable,KineticaConnection,character-method}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}}

Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbFetch}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaConnection methods}
\concept{KineticaDriver methods}
\concept{KineticaResult methods}
