% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbGetQuery,KineticaConnection,character-method}
\alias{dbGetQuery,KineticaConnection,character-method}
\title{Executes the Query and returns the resulting data.frame of n rows (or less)
When n is not provided, returns all rows available}
\usage{
\S4method{dbGetQuery}{KineticaConnection,character}(conn, statement,
  n = -1, ...)
}
\arguments{
\item{n}{Number of rows to fetch, default -1}
}
\description{
Executes the Query and returns the resulting data.frame of n rows (or less)
When n is not provided, returns all rows available
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbWriteTable(con, "tableA", data.frame(a = 1:40))
dbExistsTable(con, "tableA")
ds <- dbGetQuery(con, "tableA", 3)
show(ds)
dbRemoveTable(con, "tableA")
dbDisconnect(con)
}
}
