% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbGetQuery,KineticaConnection,character-method}
\alias{dbGetQuery,KineticaConnection,character-method}
\title{dbGetQuery()}
\usage{
\S4method{dbGetQuery}{KineticaConnection,character}(conn, statement,
  n = -1, ...)
}
\arguments{
\item{conn}{A [KineticaConnection-class] object}

\item{statement}{a character string containing SQL.}

\item{n}{Number of rows to fetch, default -1}

\item{...}{Other parameters passed on to methods.}
}
\description{
Executes the Query and returns the resulting data.frame of n rows (or less)
When n is not provided, returns all rows available
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbWriteTable(con, "tableA", data.frame(a = 1:40))
dbExistsTable(con, "tableA")
# TRUE
ds <- dbGetQuery(con, "tableA", 3)
show(ds)
    a
1   1
2   2
3   3
dbRemoveTable(con, "tableA")
dbDisconnect(con)
}
}
\seealso{
Other KineticaConnection methods: \code{\link{dbAppendTable,KineticaConnection-method}},
  \code{\link{dbCreateTable,KineticaConnection-method}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbExistsTable,KineticaConnection,character-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbIsValid,KineticaConnection-method}},
  \code{\link{dbListFields,KineticaConnection,character-method}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable,KineticaConnection,character-method}},
  \code{\link{dbRemoveTable,KineticaConnection,character-method}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable,KineticaConnection,character-method}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}}
}
\concept{KineticaConnection methods}
