% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R
\docType{methods}
\name{dbGetRowCount,KineticaResult-method}
\alias{dbGetRowCount,KineticaResult-method}
\title{dbGetRowCount()}
\usage{
\S4method{dbGetRowCount}{KineticaResult}(res, ...)
}
\arguments{
\item{res}{KineticaResult-class}

\item{...}{Other parameters passed on to methods.}
}
\description{
Returns the number of rows in the returned data set
}
\seealso{
Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbFetch}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{dbIsValid}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaResult methods}
