% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbIsValid,KineticaConnection-method}
\alias{dbIsValid,KineticaConnection-method}
\title{Checks if the connection is valid
RKinetica Connection is a configuration of url, username and password to establish
connection to Kinetica DB, configuration does not go stale when memory objects get swapped to file.
Thus RKinetica accepts stale connections as valid, and allows multiple results per connection.}
\usage{
\S4method{dbIsValid}{KineticaConnection}(dbObj, ...)
}
\description{
Checks if the connection is valid
RKinetica Connection is a configuration of url, username and password to establish
connection to Kinetica DB, configuration does not go stale when memory objects get swapped to file.
Thus RKinetica accepts stale connections as valid, and allows multiple results per connection.
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbIsValid(con)
dbDisconnect(con)
}
}
