% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbIsValid,KineticaDriver-method}
\alias{dbIsValid,KineticaDriver-method}
\title{Checks if the provided Driver object is valid}
\usage{
\S4method{dbIsValid}{KineticaDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of [KineticaDriver-class] type}
}
\description{
Checks if the provided Driver object is valid
}
\examples{
\dontrun{
drv <- Kinetica()
dbIsValid(drv)
}
}
