% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbListConnections,KineticaDriver-method}
\alias{dbListConnections,KineticaDriver-method}
\title{dbListConnections()}
\usage{
\S4method{dbListConnections}{KineticaDriver}(drv, ...)
}
\arguments{
\item{drv}{an object of [KineticaDriver-class] type}

\item{...}{Other arguments ommited in generic signature}
}
\description{
Lists all active connections in KineticaDriver's connection pool
invoking show() on KineticaConnection objects
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbGetInfo}}, \code{\link{dbIsValid}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
