% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbListObjects,KineticaConnection-method}
\alias{dbListObjects,KineticaConnection-method}
\title{#' @param conn A [KineticaConnection-class] object
#' @param ... Other parameters passed on to methods.
#' @export
setGeneric("dbListObjects",
    def = function(conn, ...) standardGeneric("dbListObjects"), valueClass = "data.frame"
)}
\usage{
\S4method{dbListObjects}{KineticaConnection}(conn, prefix = NULL, ...)
}
\arguments{
\item{prefix}{character}
}
\value{
data.frame
}
\description{
#' @param conn A [KineticaConnection-class] object
#' @param ... Other parameters passed on to methods.
#' @export
setGeneric("dbListObjects",
    def = function(conn, ...) standardGeneric("dbListObjects"), valueClass = "data.frame"
)
}
