% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbListResults,KineticaConnection-method}
\alias{dbListResults,KineticaConnection-method}
\title{Lists all active results in the current connection}
\usage{
\S4method{dbListResults}{KineticaConnection}(conn, ...)
}
\description{
Lists all active results in the current connection
}
