% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbListTables,KineticaConnection-method}
\alias{dbListTables,KineticaConnection-method}
\title{#' @export
setGeneric("dbListTables",
  def = function(conn, ...) standardGeneric("dbListTables"),
           valueClass = "character")}
\usage{
\S4method{dbListTables}{KineticaConnection}(conn, name, ...)
}
\arguments{
\item{name}{character}
}
\description{
#' @export
setGeneric("dbListTables",
  def = function(conn, ...) standardGeneric("dbListTables"),
           valueClass = "character")
}
