% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\docType{methods}
\name{dbQuoteIdentifier}
\alias{dbQuoteIdentifier}
\alias{dbQuoteIdentifier,KineticaConnection,ANY-method}
\title{dbQuoteIdentifier()}
\usage{
dbQuoteIdentifier(conn, x, ...)

\S4method{dbQuoteIdentifier}{KineticaConnection,ANY}(conn, x, ...)
}
\arguments{
\item{conn}{object [KineticaConnection-class]}

\item{x}{a character vector or SQL object}

\item{...}{Other arguments passed on to methods.}
}
\description{
Escapes the provided character vector as SQL indetifier.

Escapes the provided character vector as SQL indetifier.
}
