% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\docType{methods}
\name{dbQuoteLiteral,KineticaConnection-method}
\alias{dbQuoteLiteral,KineticaConnection-method}
\title{dbQuoteLiteral()}
\usage{
\S4method{dbQuoteLiteral}{KineticaConnection}(conn, x, ...)
}
\arguments{
\item{conn}{A subclass of [KineticaConnection-class]}

\item{x}{A vector to quote as string.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Escapes a literal if nesessary.
}
