% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\docType{methods}
\name{dbQuoteString}
\alias{dbQuoteString}
\alias{dbQuoteString,KineticaConnection,ANY-method}
\alias{dbQuoteString,KineticaConnection,character-method}
\alias{dbQuoteString,KineticaConnection,SQL-method}
\title{dbQuoteString()}
\usage{
dbQuoteString(conn, x, ...)

\S4method{dbQuoteString}{KineticaConnection,ANY}(conn, x, ...)

\S4method{dbQuoteString}{KineticaConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{KineticaConnection,SQL}(conn, x, ...)
}
\arguments{
\item{conn}{A subclass of [KineticaConnection-class], representing
an active connection to an DBMS.}

\item{x}{A character vector to quote as string.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Excapes a character (string) value.

Excapes a character (string) value.

Excapes a character (string) value.

Excapes a character (string) value.
}
