% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\name{dbRemoveTable}
\alias{dbRemoveTable}
\title{dbRemoveTable()}
\usage{
dbRemoveTable(conn, name, ...)
}
\arguments{
\item{conn}{an object of [KineticaConnection-class]}

\item{name}{character string for table name or [KineticaId-class] object}

\item{...}{Other arguments omitted in generic signature}
}
\description{
Drops the table with the given name if one exists
}
\seealso{
Other KineticaConnection methods: \code{\link{dbAppendTable}},
  \code{\link{dbCreateTable}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetInfo}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid}}, \code{\link{dbListFields}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}},
  \code{\link{transactions}}
}
\concept{KineticaConnection methods}
