% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbUnloadDriver,KineticaDriver-method}
\alias{dbUnloadDriver,KineticaDriver-method}
\title{dbUnloadDriver()}
\usage{
\S4method{dbUnloadDriver}{KineticaDriver}(drv, ...)
}
\arguments{
\item{drv}{Kinetica driver object}

\item{...}{Other arguments ommited in generic signature}
}
\description{
Unload RKinetica Driver, releasing memory from Driver connections and embedded results
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbGetInfo}}, \code{\link{dbIsValid}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
