% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\docType{methods}
\name{dbUnquoteIdentifier}
\alias{dbUnquoteIdentifier}
\alias{dbUnquoteIdentifier,KineticaConnection-method}
\title{dbUnquoteIdentifier()}
\usage{
dbUnquoteIdentifier(conn, x, ...)

\S4method{dbUnquoteIdentifier}{KineticaConnection}(conn, x, ...)
}
\arguments{
\item{conn}{object [KineticaConnection-class]}

\item{x}{a character vector or SQL object}

\item{...}{Other arguments passed on to methods.}

\item{conn}{A subclass of [KineticaConnection-class], representing
an active connection to an DBMS.}

\item{x}{An [SQL] object or character vector to unquote.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Unescapes the DB identifier.

Unescapes the DB identifier.
}
