% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaResult.R, R/kineticaSQL.R,
%   R/kineticaConnection.R
\docType{methods}
\name{dbFetch,KineticaResult-method}
\alias{dbFetch,KineticaResult-method}
\alias{dbQuoteIdentifier,KineticaConnection,ANY-method}
\alias{dbQuoteIdentifier,KineticaConnection,character-method}
\alias{dbQuoteIdentifier,KineticaConnection,SQL-method}
\alias{sqlCreateTable,KineticaConnection-method}
\title{Customised version of the generic sqlCreateTable method
(Kinetica DB uses TEMP keyword instead of TEMPORARY)}
\usage{
\S4method{dbFetch}{KineticaResult}(res, n = -1, ...)

\S4method{dbQuoteIdentifier}{KineticaConnection,ANY}(conn, x, ...)

\S4method{dbQuoteIdentifier}{KineticaConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{KineticaConnection,SQL}(conn, x, ...)

\S4method{sqlCreateTable}{KineticaConnection}(con, table, fields,
  row.names = NA, temporary = FALSE, ...)
}
\arguments{
\item{n}{maximum number of records to retrieve per fetch. Use `n = -1`}

\item{conn}{object [KineticaConnection-class]}

\item{x}{a character vector or SQL object}

\item{row.names}{a flag with logical, character or NULL value}

\item{temporary}{logical TRUE/FALSE for temporary/permanent table created}

\item{conn}{object [KineticaConnection-class]}

\item{x}{a character vector or SQL object}
}
\description{
Customised version of the generic sqlCreateTable method
(Kinetica DB uses TEMP keyword instead of TEMPORARY)
}
