% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\name{dbCanConnect}
\alias{dbCanConnect}
\title{Method allows to check if provided Kinetica Connection configuration could be used to connect
to Kinetica DB instance successfully. It takes all the arguments of dbConnect signature,
creates a connection and destroys it upon exit.}
\usage{
dbCanConnect(drv, ...)
}
\arguments{
\item{drv}{Kinetica driver object}

\item{...}{Other parameters passed on to methods.}
}
\description{
Method allows to check if provided Kinetica Connection configuration could be used to connect
to Kinetica DB instance successfully. It takes all the arguments of dbConnect signature,
creates a connection and destroys it upon exit.
}
