% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{class}
\name{KineticaDriver-class}
\alias{KineticaDriver-class}
\alias{KineticaDriver,Kinetica}
\alias{Kinetica}
\title{Class KineticaDriver}
\usage{
Kinetica()
}
\description{
This is the \pkg{RKinetica} implementation of base [DBIDriver-class] class
from \pkg{DBI}. Driver object is created with \code{Kinetica()} constructor, which
does not need arguments.
See "Usage" section for DBI methods of [DBIDriver-class] overriden by \pkg{RKinetica}.
KineticaDriver defines operations for validating connection configuration \code{dbCanConnect()},
creating connections \code{dbConnect()}, mapping data types \code{dbDataType()},
listing active connections \code{dbListConnections()}, destroying connections \code{dbDisconnect()},
etc.

Default constructor for KineticaDriver class
}
\examples{
\dontrun{
drv <- RKinetica::Kinetica()
d <- Kinetica()
}
}
\seealso{
dbConnect,dbCanConnect
}
\keyword{internal}
