% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaId.R
\docType{class}
\name{KineticaId-class}
\alias{KineticaId-class}
\alias{KineticaId}
\title{Refer to a table nested in a hierarchy (e.g. within a schema)}
\usage{
KineticaId(...)
}
\arguments{
\item{...}{Components of the hierarchy, e.g. `schema`, `table`,
or `cluster`, `catalog`, `schema`, `table`.
For more on these concepts, see
\url{http://stackoverflow.com/questions/7022755/}}
}
\description{
Objects of class `KineticaId` have a single slot `name`, which is a named
character vector.
The [dbQuoteIdentifier()] method converts `KineticaId` objects to strings.
Support for `KineticaId` objects depends on the database backend.
They can be used in the following methods as `name` argument:
- [dbCreateTable()]
- [dbAppendTable()]
- [dbReadTable()]
- [dbWriteTable()]
- [dbExistsTable()]
- [dbRemoveTable()]
Objects of this class are also returned from [dbListObjects()].
}
\examples{
KineticaId(schema = "dbo", table = "Customer")
dbQuoteIdentifier(Kinetica(), KineticaId(schema = "ki_home", table = "nyctaxi"))
KineticaId(schema = "myschema", table = "mytable")
}
