% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\name{quote_string}
\alias{quote_string}
\title{# quote_string <-
#   function(conn, x, ...) {
# #    if (is(x, "SQL")) return(x)
#     if (!is.character(x)) stop("x must be character or SQL", call. = FALSE)
#     x <- gsub("'", "''")
#     str <- paste("'", x, "'", sep = "")
# }}
\usage{
quote_string(conn, x, ...)
}
\description{
# quote_string <-
#   function(conn, x, ...) {
# #    if (is(x, "SQL")) return(x)
#     if (!is.character(x)) stop("x must be character or SQL", call. = FALSE)
#     x <- gsub("'", "''")
#     str <- paste("'", x, "'", sep = "")
# }
}
