% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKinetica-package.R
\docType{package}
\name{RKinetica}
\alias{RKinetica}
\alias{RKinetica-package}
\title{RKinetica}
\description{
RKinetica is an R package providing access to Kinetica DB through REST interface.
Kinetica DB is a  distributed, in-memory, GPU-accelerated database that utilizes a
powerful combination of CPUs and GPUs to analyze massive, complex datasets with
millisecond response times.
}
\details{
At the core of Kinetica is a vectorized, columnar, memory-first database designed for
analytical (OLAP) workloads. Kinetica automatically distributes any workload across CPUs
and GPUs for optimal results.

For more information on Kinetica DB please visit our web site \url{http://www.kinetica.com}
}
\keyword{package}
