% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{show,KineticaConnection-method}
\alias{show,KineticaConnection-method}
\title{Minimal KineticaConnection string representation
Provides url of the Kinetica DB instance}
\usage{
\S4method{show}{KineticaConnection}(object)
}
\arguments{
\item{object}{[KineticaConnection-class]}
}
\description{
Minimal KineticaConnection string representation
Provides url of the Kinetica DB instance
}
\examples{
\dontrun{
con <- dbConnect(d, url = "http://localhost:9191")
show(con)
> <KineticaConnection>
>   url: http://localhost:9191
}
}
