% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\name{SQL}
\alias{SQL}
\title{SQL()}
\usage{
SQL(x, ..., names = NULL)
}
\arguments{
\item{x}{A character vector to label as being escaped SQL.}

\item{...}{Other arguments passed on to methods. Not otherwise used.}

\item{names}{Names for the returned object, must have the same length as `x`.}
}
\value{
An object of class `SQL`.
}
\description{
Creates an escaped SQL representation of provided named character vector.
}
