% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\name{sqlAppendTable}
\alias{sqlAppendTable}
\title{sqlAppendTable}
\usage{
sqlAppendTable(con, table, values, row.names = NA, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with [dbQuoteIdentifier()].}

\item{values}{A data frame. Factors will be converted to character vectors.}

\item{row.names}{a flag with logical, character or NULL value}

\item{...}{Other parameters passed on to methods.
Character vectors will be escaped with [dbQuoteString()].}
}
\description{
Compose query to insert rows into a table
}
\details{
`sqlAppendTable()` generates a single SQL string that inserts a
data frame into an existing table. `sqlAppendTableTemplate()` generates
a template suitable for use with [dbBind()].
The default methods are ANSI SQL 99 compliant.
These methods are mostly useful for backend implementers.

The `row.names` argument must be passed explicitly in order to avoid
a compatibility warning.  The default will be changed in a later release.
}
\seealso{
Other SQL generation: \code{\link{sqlAppendTableTemplate}}
}
\concept{SQL generation}
