% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\name{sqlAppendTableTemplate}
\alias{sqlAppendTableTemplate}
\title{sqlAppendTableTemplate()}
\usage{
sqlAppendTableTemplate(con, table, values, row.names = NA,
  prefix = "?", ..., pattern = "1")
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with [dbQuoteIdentifier()].}

\item{values}{A data frame. Factors will be converted to character vectors.}

\item{row.names}{a flag with logical, character or NULL value}

\item{prefix}{Parameter prefix to use for placeholders.}

\item{...}{Other parameters passed on to methods.
Character vectors will be escaped with [dbQuoteString()].}

\item{pattern}{Parameter pattern to use for placeholders:
- `""`: no pattern
- `"1"`: position
- anything else: field name}
}
\description{
Creates a placeholder-based template
}
\seealso{
Other SQL generation: \code{\link{sqlAppendTable}}
}
\concept{SQL generation}
