% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaObject.R
\docType{methods}
\name{summary,KineticaObject-method}
\alias{summary,KineticaObject-method}
\title{summary()}
\usage{
\S4method{summary}{KineticaObject}(object, ...)
}
\arguments{
\item{object}{An object inheriting from [KineticaObject-class]}

\item{...}{Other arguments to methods.}
}
\description{
Implementation of summary() for KineticaObject signature.
}
