% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbAppendTable,KineticaConnection-method}
\alias{dbAppendTable,KineticaConnection-method}
\title{dbAppendTable()}
\usage{
\S4method{dbAppendTable}{KineticaConnection}(conn, name, value, ...,
  row.names = NULL)
}
\arguments{
\item{conn}{an object of [KineticaConnection-class]}

\item{name}{character string for table name}

\item{value}{A data frame. Factors will be converted to character vectors.}

\item{...}{Other arguments omitted in generic signature}

\item{row.names}{a flag with logical, character or NULL value}
}
\description{
Appends rows to an existing table
}
\seealso{
Other KineticaConnection methods: \code{\link{dbCreateTable,KineticaConnection-method}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbExistsTable,KineticaConnection,character-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid,KineticaConnection-method}},
  \code{\link{dbListFields,KineticaConnection,character-method}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable,KineticaConnection,ANY-method}},
  \code{\link{dbRemoveTable,KineticaConnection,character-method}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable,KineticaConnection,ANY-method}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}},
  \code{\link{transactions}}
}
\concept{KineticaConnection methods}
