% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbCanConnect,KineticaDriver-method}
\alias{dbCanConnect,KineticaDriver-method}
\title{dbCanConnect()}
\usage{
\S4method{dbCanConnect}{KineticaDriver}(drv, host = NULL, port = NULL,
  url = NULL, username = NULL, password = NULL, timeout = NULL,
  ...)
}
\arguments{
\item{drv}{Kinetica driver object}

\item{host}{character string for Kinetica DB host address}

\item{port}{integer value for Kinetica DB port}

\item{url}{character string for Kinetica DB url (protocol + host + port)}

\item{username}{character string for Kinetica DB username}

\item{password}{character string for Kinetica DB password}

\item{timeout}{integer value for Kinetica DB connection timeout}

\item{...}{Other arguments ommited in generic signature}
}
\description{
Method allows to check if provided Kinetica Connection configuration could be used to connect
to Kinetica DB instance successfully. It takes all the arguments of dbConnect signature,
creates a connection and destroys it upon exit.
}
\examples{
\dontrun{
# If you have a Kinetica DB instance running in local docker image,
# the minimum set of parameters to connect is url
dbCanConnect(Kinetica(), url = "http://localhost:9191")
# TRUE
# You can also use host and port combination
dbCanConnect(Kinetica(), host = "127.0.0.1", port = 9191)
# TRUE
# Remote instances would also require username and password.
dbCanConnect(Kinetica(), host = "127.0.0.1", port = 9191,
    username = "username", password = "Pa$$w0rd")
# TRUE
# You can obscure username and password values by calling RStudio secure dialogue box:
dbCanConnect(Kinetica(), host = "127.0.0.1", port = 9191,
             username = rstudioapi::askForPassword("Kinetica DB user"),
             password = rstudioapi::askForPassword("Kinetica DB password"))
# TRUE
}
}
\seealso{
Other KineticaDriver methods: \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbGetInfo}}, \code{\link{dbIsValid}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
