% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbDataType,KineticaConnection-method}
\alias{dbDataType,KineticaConnection-method}
\title{Redirect to driver dbDataType method for data type mapping of provided object}
\usage{
\S4method{dbDataType}{KineticaConnection}(dbObj, obj, ...)
}
\arguments{
\item{obj}{generic object}
}
\description{
Redirect to driver dbDataType method for data type mapping of provided object
}
