% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaDriver.R
\docType{methods}
\name{dbDataType,KineticaDriver-method}
\alias{dbDataType,KineticaDriver-method}
\title{dbDataType()}
\usage{
\S4method{dbDataType}{KineticaDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{KineticaDriver object}

\item{obj}{any data object or literal}

\item{...}{Other arguments omitted in generic signature}
}
\description{
Takes an R object or literal and returns a matching SQLtype
used when creating Kinetica DB table value to store it
}
\examples{
\dontrun{
dbDataType(Kinetica(), 1L)
 "INTEGER"
dbDataType(Kinetica(), 1)
 "DOUBLE"
dbDataType(Kinetica(), "Test string")
 "VARCHAR(256)"
dbDataType(Kinetica(), Sys.Date())
 "DATE"
dbDataType(Kinetica(), list())
 "BINARY"
}
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbGetInfo}}, \code{\link{dbIsValid}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
