% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbExistsTable,KineticaConnection,character-method}
\alias{dbExistsTable,KineticaConnection,character-method}
\title{dbExistsTable()}
\usage{
\S4method{dbExistsTable}{KineticaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{an object of [KineticaConnection-class]}

\item{name}{character string for table or collection name}

\item{...}{Other parameters passed on to methods.}
}
\value{
logical
}
\description{
Checks if the table exists.
Provide either table name or collection name, not both
Kinetica DB table names are unique, different collections don't allow tables with the same name.
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbCreateTable(con, "tableA", data.frame(a = 1))
dbExistsTable(con, "tableA")
# TRUE
dbExecute(con, "DROP TABLE tableA")
# 1
dbExistsTable(con, "tableA")
# FALSE
dbDisconnect(con)
}
}
\seealso{
Other KineticaConnection methods: \code{\link{dbAppendTable,KineticaConnection-method}},
  \code{\link{dbCreateTable,KineticaConnection-method}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid,KineticaConnection-method}},
  \code{\link{dbListFields,KineticaConnection,character-method}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable,KineticaConnection,ANY-method}},
  \code{\link{dbRemoveTable,KineticaConnection,character-method}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable,KineticaConnection,ANY-method}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}}
}
\concept{KineticaConnection methods}
