% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaObject.R
\name{dbIsReadOnly}
\alias{dbIsReadOnly}
\title{dbIsReadOnly()}
\usage{
dbIsReadOnly(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from [KineticaObject-class]}

\item{...}{Other arguments to methods.}
}
\value{
logical
}
\description{
Generic dbIsReadOnly() implementation.
}
