% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaObject.R, R/kineticaDriver.R,
%   R/kineticaResult.R
\docType{methods}
\name{dbIsValid}
\alias{dbIsValid}
\alias{dbIsValid,KineticaDriver-method}
\alias{dbIsValid,KineticaResult-method}
\title{dbIsValid()}
\usage{
dbIsValid(dbObj, ...)

\S4method{dbIsValid}{KineticaDriver}(dbObj, ...)

\S4method{dbIsValid}{KineticaResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from [KineticaObject-class],
i.e. [KineticaDriver-class], [KineticaConnection-class],
or a [KineticaResult-class]}

\item{...}{Other arguments to methods.}

\item{dbObj}{an object of [KineticaDriver-class] type}

\item{...}{Other arguments omitted in generic signature}

\item{dbObj}{[KineticaResult-class]}

\item{...}{Other parameters passed on to methods.}
}
\value{
logical
}
\description{
Generic dbIsValid() implementation.

Checks if the provided Driver object is valid

Checks if the KineticaResult object has been cleared
}
\examples{
\dontrun{
drv <- Kinetica()
dbIsValid(drv)
#  TRUE
}
}
\seealso{
Other KineticaDriver methods: \code{\link{dbCanConnect,KineticaDriver-method}},
  \code{\link{dbConnect,KineticaDriver-method}},
  \code{\link{dbDataType,KineticaDriver-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbListConnections,KineticaDriver-method}},
  \code{\link{dbUnloadDriver,KineticaDriver-method}},
  \code{\link{show,KineticaDriver-method}}

Other KineticaResult methods: \code{\link{dbBind,KineticaResult-method}},
  \code{\link{dbClearResult,KineticaResult-method}},
  \code{\link{dbColumnInfo}}, \code{\link{dbFetch}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetRowCount,KineticaResult-method}},
  \code{\link{dbGetRowsAffected,KineticaResult-method}},
  \code{\link{dbGetStatement,KineticaResult-method}},
  \code{\link{dbHasCompleted,KineticaResult-method}},
  \code{\link{show,KineticaDriver-method}}
}
\concept{KineticaDriver methods}
\concept{KineticaResult methods}
