% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaConnection.R
\docType{methods}
\name{dbListFields,KineticaConnection,character-method}
\alias{dbListFields,KineticaConnection,character-method}
\title{dbListFields()}
\usage{
\S4method{dbListFields}{KineticaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A [KineticaConnection-class] object}

\item{name}{a character string for table name}

\item{...}{Other parameters passed on to methods.}
}
\description{
Lists field names of a given table
}
\examples{
\dontrun{
con <- dbConnect(Kinetica(), url = "http://localhost:9191")
dbWriteTable(con, "tableA", data.frame(a = 1:40))
dbExistsTable(con, "tableA")
# TRUE
fields <- dbListFields(con, "tableA")
fields
# "a"
dbRemoveTable(con, "tableA")
dbDisconnect(con)
}
}
\seealso{
Other KineticaConnection methods: \code{\link{dbAppendTable,KineticaConnection-method}},
  \code{\link{dbCreateTable,KineticaConnection-method}},
  \code{\link{dbDataType,KineticaConnection-method}},
  \code{\link{dbDisconnect,KineticaConnection-method}},
  \code{\link{dbExecute,KineticaConnection,character-method}},
  \code{\link{dbExistsTable,KineticaConnection,character-method}},
  \code{\link{dbGetInfo}},
  \code{\link{dbGetQuery,KineticaConnection,character-method}},
  \code{\link{dbIsValid}},
  \code{\link{dbListObjects,KineticaConnection-method}},
  \code{\link{dbListResults,KineticaConnection-method}},
  \code{\link{dbListTables,KineticaConnection-method}},
  \code{\link{dbReadTable,KineticaConnection,ANY-method}},
  \code{\link{dbRemoveTable,KineticaConnection,character-method}},
  \code{\link{dbSendQuery,KineticaConnection,character-method}},
  \code{\link{dbSendStatement,KineticaConnection,character-method}},
  \code{\link{dbWriteTable,KineticaConnection,ANY-method}},
  \code{\link{show,KineticaDriver-method}},
  \code{\link{sqlCreateTable,KineticaConnection-method}},
  \code{\link{transactions}}
}
\concept{KineticaConnection methods}
