% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineticaSQL.R
\docType{methods}
\name{dbUnquoteIdentifier}
\alias{dbUnquoteIdentifier}
\alias{dbUnquoteIdentifier,KineticaConnection,SQL-method}
\alias{dbUnquoteIdentifier,KineticaConnection,ANY-method}
\title{dbUnquoteIdentifier()}
\usage{
dbUnquoteIdentifier(conn, x, ...)

\S4method{dbUnquoteIdentifier}{KineticaConnection,SQL}(conn, x, ...)

\S4method{dbUnquoteIdentifier}{KineticaConnection,ANY}(conn, x, ...)
}
\arguments{
\item{conn}{object [KineticaConnection-class]}

\item{x}{a character vector or SQL object}

\item{...}{Other arguments passed on to methods.}

\item{conn}{A subclass of [KineticaConnection-class], representing
an active connection to an DBMS.}

\item{conn}{A subclass of [KineticaConnection-class], representing
an active connection to an DBMS.}
}
\description{
Unescapes the DB identifier.

Unescapes the DB identifier.

Unescapes the DB identifier.
}
